+++ src/azure-cli/azure/cli/command_modules/rdbms/flexible_server_commands.py
@@ -88,7 +88,6 @@ def load_flexibleserver_command_table(self, _):
     with self.command_group('postgres flexible-server', postgres_flexible_servers_sdk,
                             custom_command_type=flexible_servers_custom_postgres,
                             client_factory=cf_postgres_flexible_servers) as g:
-        #g.custom_command('create', '_flexible_server_create')
         g.custom_command('create', '_flexible_server_create', table_transformer=table_transform_output)
         g.custom_command('restore', '_flexible_server_restore', supports_no_wait=True)
         g.command('start', 'start')
@@ -148,7 +147,6 @@ def load_flexibleserver_command_table(self, _):
     with self.command_group('mysql flexible-server', mysql_flexible_servers_sdk,
                             custom_command_type=flexible_servers_custom_mysql,
                             client_factory=cf_mysql_flexible_servers) as g:
-        #g.custom_command('create', '_flexible_server_create')
         g.custom_command('create', '_flexible_server_create', table_transformer=table_transform_output)
         g.custom_command('restore', '_flexible_server_restore', supports_no_wait=True)
         g.command('start', 'start')
