+++ src/azure-cli/azure/cli/command_modules/rdbms/flexible_server_custom_common.py
@@ -5,22 +5,17 @@
 
 # pylint: disable=unused-argument, line-too-long
 
-import uuid
-from msrestazure.azure_exceptions import CloudError
-from msrestazure.tools import resource_id, is_valid_resource_id, parse_resource_id  # pylint: disable=import-error
 from knack.log import get_logger
-from azure.cli.core.commands.client_factory import get_subscription_id
-from azure.cli.core.util import CLIError, sdk_no_wait
-from ._client_factory import  get_postgresql_flexible_management_client, get_mysql_flexible_management_client
 
 logger = get_logger(__name__)
 
-## Common functions imported by other providers
 
+## Common functions imported by other providers
 def _server_list_custom_func(client, resource_group_name=None):
     if resource_group_name:
         return client.list_by_resource_group(resource_group_name)
     return client.list()
+
 
 def _flexible_firewall_rule_update_custom_func(instance, start_ip_address=None, end_ip_address=None):
     if start_ip_address is not None:
